Update Gruntfile
For this new theme I want to make things easier and prettier. This
led me to update the way Grunt works with the Theme.

Right now the Gruntfile is pretty simple and just do js hintin, sass
compilation and image optimization. But this is a new structure to build
against.

I also choose to go with .js instead of .coffee, as not everybody is using
.coffee here in Mkt Virtual.