Update style base
main.scss is the new central hub for the styles in the project.
Everything is imported there. By applying this structure, developers
have a lot more power under the hood.

Some comments give a hint where each import will be placed.

Add inuit-media and inuit-buttons object to project as it is too common.

Refer to https://github.com/inuitcss/getting-started if needed